requirements.txt
# Hukuk RAG MVP Requirements

# Core ML & Vector DB
chromadb==0.4.18
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2

# LLM Integration
openai==1.6.1
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pypandoc==1.12

# Data Processing
pandas==2.1.4
numpy==1.25.2
tiktoken==0.5.2

# Web Interface
streamlit==1.29.0
streamlit-chat==0.1.1

# Utils
pyyaml==6.0.1
tqdm==4.66.1
loguru==0.7.2

# Development
jupyter==1.0.0
black==23.12.1
pytest==7.4.3